move split functions to another file
